Rename package name for hbase module
